logger/pyproject.toml
[tool.poetry]
name = "logger"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
sqlalchemy = "^2.0.25"

[tool.poetry.group.dev.dependencies]
lib = {path = "../jb-lib", develop = true}


[tool.poetry.group.test.dependencies]
pytest = "^8.2.2"
pytest-asyncio = "^0.23.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "./"